requirements.txt
uvicorn==0.30.1
fastapi==0.110.3
python-dotenv==1.0.0
crewai==0.51.1
langchain==0.2.15
langchain-aws==0.1.17
sqlalchemy==2.0.31
tiktoken==0.7.0
pydantic==2.8.2
durationpy==0.6
async-timeout==4.0.3
psycopg2-binary==2.9.9
anthropic==0.30.1
duckduckgo-search==6.3.2
chromadb==0.4.24